make Parsers a weakdep
