+++ src/InlineStrings.jl
@@ -2,8 +2,6 @@ module InlineStrings
 
 import Base: ==
 
-using Parsers
-
 export InlineString, InlineStringType, inlinestrings
 export @inline_str
 
@@ -878,69 +876,6 @@ end
     end
 end
 
-# Parsers.xparse
-Parsers.xparse(::Type{T}, buf::AbstractString, pos, len, options, ::Type{S}=T) where {T <: InlineString, S} =
-    Parsers.xparse(T, codeunits(buf), pos, len, options, S)
-
-function Parsers.xparse(::Type{T}, source::Union{AbstractVector{UInt8}, IO}, pos, len, options::Parsers.Options, ::Type{S}=T) where {T <: InlineString, S}
-    res = Parsers.xparse(String, source, pos, len, options, PosLen)
-    code = res.code
-    overflowed = false
-    poslen = res.val
-    if !Parsers.valueok(code) || Parsers.sentinel(code)
-        x = T()
-    else
-        poslen = res.val
-        if T === InlineString1
-            if poslen.len != 1
-                overflowed = true
-                x = T()
-            else
-                Parsers.fastseek!(source, poslen.pos)
-                x = InlineString1(Parsers.peekbyte(source, poslen.pos))
-                Parsers.fastseek!(source, pos + res.tlen - 1)
-            end
-        elseif Parsers.escapedstring(code) || !(source isa AbstractVector{UInt8})
-            if poslen.len > (sizeof(T) - 1)
-                overflowed = true
-                x = T()
-            else
-                # manually build up InlineString
-                i = poslen.pos
-                maxi = i + poslen.len
-                x = T()
-                Parsers.fastseek!(source, i - 1)
-                while i < maxi
-                    b = Parsers.peekbyte(source, i)
-                    if b == options.e
-                        i += 1
-                        Parsers.incr!(source)
-                        b = Parsers.peekbyte(source, i)
-                    end
-                    x, overflowed = addcodeunit(x, b)
-                    i += 1
-                    Parsers.incr!(source)
-                end
-                Parsers.fastseek!(source, maxi)
-            end
-        else
-            vlen = poslen.len
-            if vlen > (sizeof(T) - 1)
-                # @show T, vlen, sizeof(T)
-                overflowed = true
-                x = T()
-            else
-                # @show poslen.pos, vlen
-                x = T(source, poslen.pos, vlen)
-            end
-        end
-    end
-    if overflowed
-        code |= Parsers.OVERFLOW
-    end
-    return Parsers.Result{S}(code, res.tlen, x)
-end
-
 ## InlineString sorting
 using Base.Sort, Base.Order
 
@@ -1137,5 +1072,9 @@ end
 Base.Broadcast.broadcasted(::Type{InlineString}, A::AbstractArray) = inlinestrings(A)
 Base.map(::Type{InlineString}, A::AbstractArray) = inlinestrings(A)
 Base.collect(::Type{InlineString}, A::AbstractArray) = inlinestrings(A)
+
+if !isdefined(Base, :get_extension)
+    include("../ext/ParsersExt.jl")
+end
 
 end # module
